safari-600.7-branch/LayoutTests/inspector-obsolete/styles/get-set-stylesheet-text-expected.txt

Tests that WebInspector.CSSStyleSheet methods work as expected.

Inspect Me


Running: testSetText
=== Original stylesheet text: ===
body {
    font-size: 12px;
    foo: bar !

    important /* !important property not recognized by WebCore */;
    height: 100% !important;
    colorz: "!important";
    important:!important;
    color: "badcolor" ! important /* good property with strange value */;
}

@media {
    /* @media rule with an empty media list */
}

/* comment before selector */body.main1/* comment after selector */{/* comment */color: #F00BAA;zoo:moo /* not an !important unrecognized property */}/* comment */

body.main2{background: green /* value !important comment */ !important /* no semicolon, very !important */}


Running: testNewElementStyles
=== Matched rules for h1 after setText() ===

rule
h1: [user-agent]
['display':'block'] @[undefined-undefined] style
['font-size':'2em'] @[undefined-undefined] style
['-webkit-margin-before':'0.67em'] @[undefined-undefined] style
['-webkit-margin-after':'0.67em'] @[undefined-undefined] style
['-webkit-margin-start':'0px'] @[undefined-undefined] style
['-webkit-margin-end':'0px'] @[undefined-undefined] style
['font-weight':'bold'] @[undefined-undefined] style

rule
h1: [regular]
['COLOR':'Red'] @[0:5-0:16] active
